% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ots_cache.R
\name{ots_cache}
\alias{ots_cache}
\title{Caching wrapper to reduce API calls (internal)}
\usage{
ots_cache(use_cache, file, ...)
}
\arguments{
\item{use_cache}{Logical to save and load from cache. If \code{TRUE}, the results will be cached in memory
if \code{file} is \code{NULL} or on disk if `file` is not \code{NULL}.}

\item{file}{Character with the full file path to save the data.}

\item{...}{Additional parameters inherited from \code{ots_create_tidy_data()}.}
}
\description{
Eases saving the data downloaded from \code{api.tradestatistics.io}
and prevents \code{ots_read_from_api()} from downloading the same twice.
}
\keyword{internal}
