% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatistics-package.R
\docType{data}
\name{ots_commodities_short}
\alias{ots_commodities_short}
\title{OTS Commodities Short}
\format{
A data frame with 1,225 observations on the following 2 variables
\describe{
  \item{\code{commodity_code}}{HS four digits commodity code (e.g. 0101)}
  \item{\code{commodity_fullname_english}}{HS four digits commodity names (e.g. 'Horses, asses, mules and hinnies; live')}
}
}
\source{
Open Trade Statistics
}
\usage{
ots_commodities_short
}
\description{
Official commodity names from the Harmonized System rev 2012
(HS12, four digits detail).
}
\keyword{datasets}
