\name{ots_commodities_shortnames}
\alias{ots_commodities_shortnames}
\title{A table of unofficial shortened product names}
\docType{data}
\description{
Provides shortened product names obtained from The Center for International Development at Harvard University.
This data corresponds to a modified version of unofficial short names created for 4 digits HS codes.
}
\usage{ots_commodities_shortnames}
\format{
  A data frame with 1222 observations on the following 2 variables.
  \describe{
    \item{\code{commodity_code}}{Code of every commodity (e.g. 0101)}
    \item{\code{commodity_shortname_english}}{Unofficial commodity names (e.g. Horses)}
  }
}
\examples{
ots_commodities_shortnames
}
\keyword{datasets}
