% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ots_strings_processing.R
\name{ots_commodity_community}
\alias{ots_commodity_community}
\title{String matching of unofficial commodity community codes according to
the Center for International Development at Harvard University nomenclature}
\usage{
ots_commodity_community(community = NULL)
}
\arguments{
\item{community}{A text string such as "animals", or "FOODSTUFFS".}
}
\value{
A tibble with all possible matches (no uppercase distinction)
showing the section name and section code
}
\description{
Takes a text string and searches within the
package data for all matching communities in the context of valid API
communities
}
\examples{
ots_commodity_community(community = "Animal")
ots_commodity_community(community = "  Animals")
ots_commodity_community(community = "FABRIC ")
}
\keyword{functions}
