\name{ots_communities}
\alias{ots_communities}
\title{A table of sections codes, names and colours associated to products}
\docType{data}
\description{
Provides section names from the Harmonized System.
}
\usage{ots_communities}
\format{
  A data frame with 1242 observations on the following 4 variables.
  \describe{
    \item{\code{commodity_code}}{Commodity code (e.g. 0101)}
    \item{\code{community_code}}{Community code (e.g. 01)}
    \item{\code{community_name}}{Community name (e.g. Animal Products)}
    \item{\code{community_color}}{Community color in hex code (e.g. #74c0e2)}
  }
}
\examples{
ots_communities
}
\keyword{datasets}
