% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatistics-package.R
\docType{data}
\name{ots_countries}
\alias{ots_countries}
\title{OTS Countries}
\format{
A data frame with 308 observations on the following 5 variables
\describe{
  \item{\code{country}}{Official name (e.g. United Kingdom)}
  \item{\code{iso3}}{ISO-3 code (e.g. "GBR")}
  \item{\code{dynamic_code}}{Deambiguated ISO-3 code (e.g. "DEU.X" for East Germany)}
  \item{\code{continent}}{Corresponding continent (e.g., Europe)}
  \item{\code{colour}}{Assigned colour by continen (e.g., '#d1a1bc')}
}
}
\source{
Derived from USITC
}
\usage{
ots_countries
}
\description{
Official country names, ISO-3 codes, continent and EU membership.
}
\keyword{datasets}
