% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ots_create_tidy_data.R
\name{ots_create_tidy_data}
\alias{ots_create_tidy_data}
\title{Downloads and processes the data from the API to return a human-readable tibble}
\usage{
ots_create_tidy_data(
  years = 2020,
  importers = "all",
  exporters = "all",
  sectors = "all",
  industries = "all",
  table = "itpde_imp",
  max_attempts = 5,
  use_cache = FALSE,
  file = NULL
)
}
\arguments{
\item{years}{Year contained within the years specified in
api.tradestatistics.io/year_range (e.g., \code{2002} or \code{seq(2002, 2010, 2)}).
Default set to \code{2019}.}

\item{importers}{Importers (e.g. \code{"chl"}, \code{"Chile"} or
\code{c("chl", "Peru")}). Default set to \code{"all"}.}

\item{exporters}{Exporters (e.g. \code{"chl"}, \code{"Chile"} or
\code{c("chl", "Peru")}). Default set to \code{"all"}.}

\item{sectors}{Sectors (e.g. \code{"Agriculture"}, \code{1} or search matches for \code{"agri"})
to filter sectors. Default set to \code{"all"}.}

\item{industries}{Industries (e.g. \code{"Cereal products"}, \code{5} or search matches for \code{"cereal"})
to filter industries. Default set to \code{"all"}.}

\item{table}{Character string to select the table to obtain the data.
Default set to \code{itpde_imp} (aggregate trade by importer-year).
Run \code{ots_tables} in case of doubt.}

\item{max_attempts}{How many times to try to download data in case the
API or the internet connection fails when obtaining data. Default set
to \code{5}.}

\item{use_cache}{Logical to save and load from cache. If \code{TRUE}, the results will be cached in memory
if \code{file} is \code{NULL} or on disk if `file` is not \code{NULL}. Default set to \code{FALSE}.}

\item{file}{Optional character with the full file path to save the data. Default set to \code{NULL}.}
}
\value{
A tibble that describes bilateral trade metrics (imports,
exports, trade balance and relevant metrics
such as exports growth w/r to last year) between a \code{reporter}
and \code{partner} country.
}
\description{
Accesses \code{api.tradestatistics.io} and
performs different API calls to transform and return tidy data.
}
\examples{
\dontrun{
# The next examples can take more than 5 seconds to compute,
# so these are just shown without evaluation according to CRAN rules

# Run `ots_countries` to display the full table of countries
# Run `ots_sectors` to display the full table of sectors

# Agricultural trade (sector)
ots_create_tidy_data(years = 2020, sectors = "Agriculture",
 table = "itpde_sec")

# Cereal products trade (industry-importer)
ots_create_tidy_data(years = 2020, importers = "chl",
 industries = "Cereal products", table = "itpde_imp_ind")

# Aggregate trade (bilateral)
ots_create_tidy_data(years = 2020, importers = "chl",
 exporters = c("arg", "bra"), table = "itpde_imp_exp")
}
}
\keyword{functions}
