% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ots_create_tidy_data.R
\name{ots_create_tidy_data_unmemoised}
\alias{ots_create_tidy_data_unmemoised}
\title{Downloads and processes the data from the API to return a human-readable tibble (unmemoised, internal)}
\usage{
ots_create_tidy_data_unmemoised(
  years,
  reporters,
  partners,
  products,
  table,
  max_attempts,
  include_shortnames,
  include_communities,
  use_localhost
)
}
\description{
A separation of the downloading part in \code{ots_create_tidy_data()} for making caching optional.
}
\keyword{internal}
