% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatistics-package.R
\docType{data}
\name{ots_distances}
\alias{ots_distances}
\title{OTS Distances}
\format{
A data frame with 22,791 rows and 8 variables
\itemize{
  \item{\code{country1}}{First ISO-3 code in the dyad (alphabetical order)}
  \item{\code{country2}}{Second ISO-3 code in the dyad (alphabetical order)}
  \item{\code{dist}}{Distance between most populated cities (in kilometers)}
  \item{\code{distcap}}{Distance between capitals (in kilometers)}
  \item{\code{colony}}{Variable coded as 1 when the two countries are or were in a colonial relation}
  \item{\code{comlang_ethno}}{Variable coded as 1 when the two countries have at least 9\% of their population speaking the same language}
  \item{\code{comlang_off}}{Variable coded as 1 when the two countries share the same official language}
  \item{\code{contig}}{Variable coded as 1 when the two countries are next to each other and 0 otherwise}
}
}
\source{
Adapted from CEPII
}
\usage{
ots_distances
}
\description{
Distance between countries, alongside colonial relation, common language,
and continuity.
}
\keyword{datasets}
