% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatistics-package.R
\docType{data}
\name{ots_gdp_deflator}
\alias{ots_gdp_deflator}
\title{GDP Deflator}
\format{
A data frame with 7,238 observations on the following 4 variables
\itemize{
  \item{\code{country_iso}}{ISO code of the country (e.g. "chl" means Chile)}
  \item{\code{from}}{Integer values in the range 1980-2018}
  \item{\code{to}}{Integer values in the range 1981-2019}
  \item{\code{gdp_deflator}}{Numeric value expressed as one plus 1-year deflator}
}
}
\source{
Open Trade Statistics
}
\usage{
ots_gdp_deflator
}
\description{
A table with the World's weigthed GDP deflator inflation since 2000.
Provides year to year GDP deflator values for every country listed in the 
World Bank database. This data is ready to be applied as a conversion rate 
to express dollars of year Y1 as dollars of year Y2.
For countries not listed in the World Bank database, rows labelled as "wld" 
are available, which were computed as the weighted median for each year using 
the GDP of listed countries for each year expressed as constant dollars of 
the year 2010.
This dataset is provided to be used with \code{ots_gdp_deflator_adjustment} 
that converts units forwards and backwards in time.
}
\keyword{datasets}
