\name{ots_groups}
\alias{ots_groups}
\title{A table of group codes and names associated to products}
\docType{data}
\description{
Provides section names from the Harmonized System.
}
\usage{ots_groups}
\format{
  A data frame with 97 observations on the following 2 variables.
  \describe{
    \item{\code{group_code}}{Group code (e.g. 01)}
    \item{\code{group_fullname_english }}{Group name (e.g. Animals; live  )}
  }
}
\examples{
ots_groups
}
\keyword{datasets}
