% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatistics-package.R
\docType{data}
\name{ots_industries}
\alias{ots_industries}
\title{OTS Industries}
\format{
A data frame with 170 observations on the following 2 variables
\describe{
  \item{\code{industry_descr}}{Industry name (e.g. 'Wheat')}
  \item{\code{industry_id}}{Industry code (e.g. '1')}
}
}
\source{
Derived from USITC
}
\usage{
ots_industries
}
\description{
Industry IDs.
}
\keyword{datasets}
