% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ots_strings_processing.R
\name{ots_industry_code}
\alias{ots_industry_code}
\title{String matching of industry names and IDs}
\usage{
ots_industry_code(industry = NULL)
}
\arguments{
\item{industry}{A text string such as "Wheat", "CEREAL" or "oilseed".}
}
\value{
A data frame with all possible matches (no uppercase distinction)
showing the industry name and industry ID.
}
\description{
Takes a text string and searches within the
package data for all matching industry IDs in the context of valid API
industry codes.
}
\examples{
ots_industry_code("Wheat")
ots_industry_code("cereal")
}
\keyword{functions}
