% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ots_strings_processing.R
\name{ots_product_community}
\alias{ots_product_community}
\title{String matching of unofficial product community names and product community
codes}
\usage{
ots_product_community(productcommunity = NULL)
}
\arguments{
\item{productcommunity}{A text string such as "animals", or "FOODSTUFFS".}
}
\value{
A tibble with all possible matches (no uppercase distinction)
showing the community name and community code
}
\description{
Takes a text string and searches within the
package data for all matching product communities in the context of valid API
product communities
}
\examples{
ots_product_community(productcommunity = "  Animals")
ots_product_community(productcommunity = "FABRIC ")
}
\keyword{functions}
