% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ots_strings_processing.R
\name{ots_product_section}
\alias{ots_product_section}
\title{String matching of unofficial product section names and product section
codes}
\usage{
ots_product_section(productsection = NULL)
}
\arguments{
\item{productsection}{A text string such as "animals", or "FOODSTUFFS".}
}
\value{
A tibble with all possible matches (no uppercase distinction)
showing the section name and section code
}
\description{
Takes a text string and searches within the
package data for all matching product communities in the context of valid API
product communities
}
\examples{
ots_product_section(productsection = "  Animals")
ots_product_section(productsection = "FABRIC ")
}
\keyword{functions}
