\name{ots_product_shortnames}
\alias{ots_product_shortnames}
\title{A table of unofficial shortened product names}
\docType{data}
\description{
Provides shortened product names obtained from The Observatory of Economic Complexity.
This data corresponds to a modified version of unofficial short names created for 4 digits HS codes.
}
\usage{ots_products}
\format{
  A data frame with 1222 observations on the following 4 variables.
  \describe{
    \item{\code{product_code}}{Code of every product (e.g. 0101)}
    \item{\code{product_shortname_english}}{Unofficial product names (e.g. Horses)}
  }
}
\examples{
ots_products
}
\keyword{datasets}
