\name{ots_sections_colors}
\alias{ots_sections_colors}
\title{A table of official section names from the Harmonized System rev 2012
 (HS12, also known as H4) and unofficial colors to ease visualization}
\docType{data}
\description{
Provides official section codes and names taken from the United Nations official
sources.
}
\usage{ots_sections_colors}
\format{
  A data frame with 22 observations on the following 2 variables.
  \describe{
    \item{\code{section_code }}{Section code (e.g. '01')}
    \item{\code{section_name}}{Section hex color (e.g. 'Live animals and animal products')}
  }
}
\examples{
ots_sections_colors
}
\keyword{datasets}
