\name{ots_sections_names}
\alias{ots_sections_names}
\title{A table of sections codes and official names associated to products}
\docType{data}
\description{
Provides section names from the Harmonized System.
}
\usage{ots_sections_names}
\format{
  A data frame with 2 observations on the following 2 variables.
  \describe{
    \item{\code{section_code}}{Section code (e.g. 01)}
    \item{\code{section_fullname_english}}{Section name (e.g. Live Animals; Animal Products)}
  }
}
\examples{
ots_sections_names
}
\keyword{datasets}
