\name{ots_sections_shortnames}
\alias{ots_sections_shortnames}
\title{A table of sections codes and shortened names associated to products}
\docType{data}
\description{
Provides shortened section names from The Atlas of Economic Complexity (Harvard), which are
simplified names after official HS sections but keeping the same codes.
}
\usage{ots_sections}
\format{
  A data frame with 22 observations on the following 2 variables.
  \describe{
    \item{\code{section_code}}{Section code (e.g. 01)}
    \item{\code{section_shortname_english}}{Section name (e.g. Animal Products)}
  }
}
\examples{
ots_sections_shortnames
}
\keyword{datasets}
