% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ots_strings_processing.R
\name{ots_sector_code}
\alias{ots_sector_code}
\title{String matching of broad sector names and IDs}
\usage{
ots_sector_code(sector = NULL)
}
\arguments{
\item{sector}{A text string such as "Agriculture", "MINING" or "serv".}
}
\value{
A data frame with all possible matches (no uppercase distinction)
showing the sector name and sector ID.
}
\description{
Takes a text string and searches within the
package data for all matching sector IDs in the context of valid API
sector codes.
}
\examples{
ots_sector_code("Agriculture")
ots_sector_code("mining")
}
\keyword{functions}
