% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatistics-package.R
\docType{data}
\name{ots_sectors}
\alias{ots_sectors}
\title{OTS Sectors}
\format{
A data frame with 4 observations on the following 2 variables
\describe{
  \item{\code{broad_sector}}{Sector name (e.g. 'Agriculture')}
  \item{\code{broad_sector_id}}{Sector code (e.g. '1')}
  \item{\code{colour}}{Sector colour (e.g., '#74c0e2')}
}
}
\source{
Derived from USITC
}
\usage{
ots_sectors
}
\description{
Broad sector IDs.
}
\keyword{datasets}
