% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatistics-package.R
\docType{data}
\name{ots_tables}
\alias{ots_tables}
\title{OTS Tables}
\format{
A data frame with 16 rows and 3 variables
\itemize{
  \item{\code{table}}{Table name}
  \item{\code{description}}{Description of table contents}
  \item{\code{source}}{Source for the data (OTS tables are processed after UN Comtrade raw data)}
}
}
\source{
Open Trade Statistics
}
\usage{
ots_tables
}
\description{
A table describing existing API tables with both description and source.
This data is used by the functions provided within this package to validate
user parameters.
}
\keyword{datasets}
