% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ots_trade_data.R
\name{ots_trade_data}
\alias{ots_trade_data}
\title{Downloads and processes the data from the API}
\usage{
ots_trade_data(years = NULL, reporter = NULL, partner = NULL,
  commodity_code_length = 4, table = "yrpc", max_attempts = 5)
}
\arguments{
\item{years}{Numeric value greater or equal to 1962 and lower of equal
to 2016. Default set to \code{NULL}.}

\item{reporter}{ISO code for country of reporter (e.g. \code{chl} for
Chile). Default set to \code{NULL}.
Run \code{countries} in case of doubt.}

\item{partner}{ISO code for country of partner (e.g. \code{chn} for
China). Default set to \code{NULL}.
Run \code{countries} in case of doubt.}

\item{commodity_code_length}{Character string to indicate the granularity level on commodities.
Default set to \code{4} (it can also take the values \code{6} or
\code{all}).}

\item{table}{Character string to select the table to obtain the data.
Default set to \code{yrpc} (Year - Reporter - Partner - Commodity).
Run \code{tables} in case of doubt.}

\item{max_attempts}{How many times to try to download data in case the
API or the internet connection fails when obtaining data. Default set
to \code{5}.}
}
\value{
A tibble that describes bilateral trade metrics (imports,
exports, trade balance and relevant metrics
such as exports growth w/r to last year) between an \code{reporter}
and \code{partner} country.
}
\description{
This function accesses \code{api.tradestatistics.io} and
performs different API calls to return tidy data. and data transforming.
}
\examples{
\dontrun{
# The next example can take more than 5 seconds to compute,
# so these are just shown without evaluation according to CRAN rules

# Run `countries` to display the full table of countries

# What does Chile export to China?
ots_trade_data(reporter = "chl", partner = "chn", years = 2015)
}
}
\keyword{functions}
