% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_logic.R, R/ots_cache_logic.R
\name{ots_with_cache}
\alias{ots_with_cache}
\title{Caching wrapper to reduce API calls (internal function)}
\usage{
ots_with_cache(use_cache, file, memoised, unmemoised, ...)

ots_with_cache(use_cache, file, memoised, unmemoised, ...)
}
\arguments{
\item{use_cache}{Logical to save and load from cache. If \code{TRUE}, the results will be cached in memory
if \code{file} is \code{NULL} or on disk if `file` is not \code{NULL}.}

\item{file}{Character with the full file path to save the data.}

\item{memoised}{a memorised function (e.g. \code{ots_create_tidy_data_memoised()}).}

\item{unmemoised}{a memorised function (e.g. \code{ots_create_tidy_data_unmemoised()}).}

\item{...}{additional parameters inherited from \code{ots_create_tidy_data()}.}
}
\description{
This function eases saving the data downloaded from \code{api.tradestatistics.io}
and prevents \code{ots_read_from_api()} from downloading the same twice.

This function eases saving the data downloaded from \code{api.tradestatistics.io}
and prevents \code{ots_read_from_api()} from downloading the same twice.
}
\keyword{internal}
