% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_code.R
\name{product_code}
\alias{product_code}
\title{Obtain a table of official product names and Harmonized System (HS) codes 
according to the United Nations nomenclature}
\usage{
product_code(productname = NULL)
}
\arguments{
\item{productname}{A text string such as "Animals", "COPPER" or "fruits".}
}
\value{
A tibble with all possible matches (no uppercase distinction) 
showing the product name, product code and corresponding trade 
classification (e.g. HS92 or SITC)
}
\description{
This function takes a text string and searches within the 
package data for all matching product codes in the context of valid API 
product codes.
}
\examples{
product_code("animals")
product_code("fruits")
}
\keyword{functions}
