\name{tables}
\alias{tables}
\title{Available tables in the API}
\docType{data}
\description{
Provides a detail of valid queries by table.
}
\usage{tables}
\format{
  A data frame with 10 observations on the following 3 variables.
  \describe{
    \item{\code{table}}{Table name}
    \item{\code{description}}{Description of table contents}
    \item{\code{example_query}}{Examples of valid queries}
  }
}
\examples{
products
}
\keyword{datasets}
