% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatistics-package.R
\docType{package}
\name{tradestatistics-package}
\alias{tradestatistics}
\alias{tradestatistics-package}
\title{tradestatistics: Open Trade Statistics API Wrapper and Utility Program}
\description{
Access Open Trade Statistics API from R to download international trade data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ropensci.github.io/tradestatistics}
  \item Report bugs at \url{https://github.com/ropensci/tradestatistics/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas \email{mvargas@dcc.uchile.cl} (0000-0003-1017-7574) [copyright holder]

Other contributors:
\itemize{
  \item Joshua Kunst [contributor]
  \item Emily Riederer [reviewer]
  \item Mark Padgham (0000-0003-1017-7574) [reviewer]
  \item Amanda Dobbyn [reviewer]
  \item Jorge Cimentada [reviewer]
  \item  United Nations International Trade Statistics Database [data contributor]
  \item  Center for International Development at Harvard University [data contributor]
  \item  The Observatory of Economic Complexity [data contributor]
}

}
\keyword{internal}
