% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatistics-package.R
\docType{data}
\name{ots_commodities}
\alias{ots_commodities}
\title{OTS Commodities}
\format{
A data frame with 5,304 observations on the following 4 variables
\itemize{
  \item{\code{commodity_code}}{Code of every commodity (e.g. 010110)}
  \item{\code{commodity_fullname_english}}{HS commodity names (e.g. 'Horses, asses, mules and hinnies; live, pure-bred breeding animals')}
  \item{\code{section_code}}{Section code (e.g. 01)}
  \item{\code{section_fullname_english }}{Section name (e.g. 'Live animals and animal products')}
}
}
\source{
Open Trade Statistics
}
\usage{
ots_commodities
}
\description{
A table of official commodity names from the Harmonized System rev 2012
(HS12, also known as H4). Provides official commodity and section codes and 
names at six (6) digits detail taken from the United Nations official 
sources.
This data is used by the functions provided within this package to complement
the data obtained from the API.
}
\keyword{datasets}
