% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatistics-package.R
\docType{data}
\name{ots_commodities_short}
\alias{ots_commodities_short}
\title{OTS Commodities Short}
\format{
A data frame with 1,225 observations on the following 2 variables
\itemize{
  \item{\code{commodity_code}}{Code of every commodity (e.g. 010110)}
  \item{\code{commodity_fullname_english}}{HS commodity names (e.g. 'Horses, asses, mules and hinnies; live, pure-bred breeding animals')}
  \item{\code{section_code}}{Section code (e.g. 01)}
  \item{\code{section_fullname_english }}{Section name (e.g. 'Live animals and animal products')}
}
}
\source{
Open Trade Statistics
}
\usage{
ots_commodities_short
}
\description{
Official commodity names from the Harmonized System rev 2012
(HS12, four digits detail).
}
\keyword{datasets}
