\name{ots_sections}
\alias{ots_sections}
\title{A table of sections codes, names and colours associated to products}
\docType{data}
\description{
Provides sections from The Atlas of Economic Complexity (Harvard), which are
simplified names after official HS sections but keeping the same codes.
}
\usage{ots_sections}
\format{
  A data frame with 1222 observations on the following 4 variables.
  \describe{
    \item{\code{product_code}}{Code of every product (e.g. 0101)}
    \item{\code{section_code}}{Section code (e.g. 01)}
    \item{\code{section_name}}{Section name (e.g. Animal Products)}
    \item{\code{section_color}}{Colour hex code (e.g. #74c0e2)}
  }
}
\examples{
ots_sections
}
\keyword{datasets}
