\name{ots_communities}
\alias{ots_communities}
\title{A table of communities codes, names and colours associated to products}
\docType{data}
\description{
Provides communities from The Atlas of Economic Complexity (Harvard), and those
can be useful for grouping into broader categories than those the official HS groups.
}
\usage{ots_communities}
\format{
  A data frame with 6275 observations on the following 4 variables.
  \describe{
    \item{\code{product_code}}{Code of every product (e.g. 0101)}
    \item{\code{community_code}}{Community code (e.g. 01)}
    \item{\code{community_name}}{Community name (e.g. Animal Products)}
    \item{\code{community_color}}{Colour hex code (e.g. #74c0e2)}
  }
}
\examples{
ots_communities
}
\keyword{datasets}
