% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatistics-package.R
\docType{data}
\name{ots_countries}
\alias{ots_countries}
\title{OTS Countries}
\format{
A data frame with 275 observations on the following 5 variables
\describe{
  \item{\code{country_iso}}{ISO-3 code of the country (e.g. "GBR")}
  \item{\code{country_name}}{Country name (e.g. United Kingdom)}
  \item{\code{continent_id}}{Numeric id of the continent where the country belongs to (e.g., 4)}
  \item{\code{continent_name}}{Continent where the country belongs to (e.g., Europe)}
  \item{\code{continent_color}}{Color assigned to each continent for visualization purposes (e.g., '#8abdb6')}
}
}
\source{
Open Trade Statistics
}
\usage{
ots_countries
}
\description{
Official country names, ISO-3 codes, continent and EU membership.
}
\keyword{datasets}
