% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatistics-package.R
\docType{data}
\name{ots_countries}
\alias{ots_countries}
\title{OTS Countries}
\format{
A data frame with 275 observations on the following 5 variables
\describe{
  \item{\code{country_iso}}{ISO-3 code of the country (e.g. "deu" means Germany)}
  \item{\code{country_name_english}}{Country name (e.g. Germany)}
  \item{\code{country_fullname_english}}{Country name with indications (e.g. Germany as "Germany (former Federal Republic of Germany until 1990)")}
  \item{\code{continent_name_english}}{Continent where the country belongs to (e.g., Europe)}
  \item{\code{continent_id}}{Numeric id of the continent where the country belongs to (e.g., 5)}
}
}
\source{
Open Trade Statistics
}
\usage{
ots_countries
}
\description{
Official country names, ISO-3 codes, continent and EU membership.
}
\keyword{datasets}
