% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatistics-package.R
\docType{data}
\name{ots_countries}
\alias{ots_countries}
\title{OTS Countries}
\format{
A data frame with 264 observations on the following 6 variables
\itemize{
  \item{\code{country_iso}}{ISO code of the country (e.g. "chl" means Chile)}
  \item{\code{country_name_english}}{Country name (e.g. Germany)}
  \item{\code{country_fullname_english}}{Country name with indications (e.g. Germany (former Federal Republic of Germany until 1990))}
  \item{\code{continent_id}}{Numeric id of the continent where the country belongs to}
  \item{\code{continent_name_english}}{Continent where the country belongs to}
  \item{\code{eu28_member}}{Dummy variable such that 1 means "belongs to EU-28 group" and 0 otherwise}
}
}
\source{
Open Trade Statistics
}
\usage{
ots_countries
}
\description{
A table of official country names, ISO-3 codes and other metadata. The
availability is limited to countries with records in the OTS database.
This data is used by the functions provided within this package to complement
the data obtained from the API.
}
\keyword{datasets}
