% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatistics-package.R
\docType{data}
\name{ots_countries_colors}
\alias{ots_countries_colors}
\title{OTS Countries Colors}
\format{
A data frame with 276 rows and 3 variables
\itemize{
  \item{\code{continent_id}}{Numeric id of the continent where the country belongs to}
  \item{\code{country_iso}}{ISO code of the country (e.g. "chl" means Chile)}
  \item{\code{country_color}}{Section hex color (e.g. '#D05555')}
}
}
\source{
Open Trade Statistics
}
\usage{
ots_countries_colors
}
\description{
Unofficial colors to ease visualization for countries.
}
\keyword{datasets}
