% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ots_country_code.R
\name{ots_country_code}
\alias{ots_country_code}
\title{String matching of official country names and ISO-3 codes according to
the United Nations nomenclature}
\usage{
ots_country_code(countryname = NULL)
}
\arguments{
\item{countryname}{A text string such as "Chile", "CHILE" or "CHL".}
}
\value{
A single character if there is a exact match (e.g.
\code{ots_country_code("Chile")}) or a tibble in case of multiple matches
(e.g. \code{ots_country_code("Germany")})
}
\description{
This function takes a text string and searches within the
package data for a country code in the context of valid API country codes.
}
\examples{
# Single match with no replacement
ots_country_code("Chile")

# Single match with replacement
ots_country_code("America")

# Double match with no replacement
ots_country_code("Germany")
}
\keyword{functions}
