% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ots_create_tidy_data.R
\name{ots_create_tidy_data}
\alias{ots_create_tidy_data}
\title{Downloads and processes the data from the API to return a human-readable tibble}
\usage{
ots_create_tidy_data(years = NULL, reporter = NULL, partner = NULL,
  product_code = "all", product_code_length = 4, table = "yrpc",
  max_attempts = 5)
}
\arguments{
\item{years}{Numeric value greater or equal to 1962 and lower of equal
to 2016. Default set to \code{NULL}.}

\item{reporter}{ISO code for country of reporter (e.g. \code{chl} for
Chile). Default set to \code{NULL}.
Run \code{countries} in case of doubt.}

\item{partner}{ISO code for country of partner (e.g. \code{chn} for
China). Default set to \code{NULL}.
Run \code{countries} in case of doubt.}

\item{product_code}{Character string (e.g. \code{0101} or \code{01}) to filter products.
Default set to \code{"all"}.}

\item{product_code_length}{Character string to indicate the granularity level on products
Default set to \code{4} (it can also take the values \code{6} or
\code{all}).}

\item{table}{Character string to select the table to obtain the data.
Default set to \code{yrpc} (Year - Reporter - Partner - Product Code).
Run \code{tables} in case of doubt.}

\item{max_attempts}{How many times to try to download data in case the
API or the internet connection fails when obtaining data. Default set
to \code{5}.}
}
\value{
A tibble that describes bilateral trade metrics (imports,
exports, trade balance and relevant metrics
such as exports growth w/r to last year) between an \code{reporter}
and \code{partner} country.
}
\description{
This function accesses \code{api.tradestatistics.io} and
performs different API calls to return tidy data. and data transforming.
}
\examples{
\dontrun{
# The next example can take more than 5 seconds to compute,
# so these are just shown without evaluation according to CRAN rules

# Run `countries` to display the full table of countries

# What does Chile export to China? (2015)
ots_create_tidy_data(years = 2015, reporter = "chl", partner = "chn")

# What can we say about Horses exported by Chile? (1980)
ots_create_tidy_data(years = 1980, product_code = "0101", table = "yc")
ots_create_tidy_data(years = 1980, reporter = "chl", product_code = "0101", table = "yrc")
ots_create_tidy_data(years = 1980, reporter = "chl", partner = "arg", product_code = "0101", 
table = "yrpc")
}
}
\keyword{functions}
