% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ots_create_tidy_data.R
\name{ots_create_tidy_data_memoised}
\alias{ots_create_tidy_data_memoised}
\title{Downloads and processes the data from the API to return a human-readable tibble (memoised, internal)}
\usage{
ots_create_tidy_data_memoised(
  years = 2018,
  reporters = "usa",
  partners = "all",
  products = "all",
  sections = "all",
  groups = "all",
  table = "yr",
  max_attempts = 5,
  use_localhost = FALSE
)
}
\description{
A composition of \code{ots_create_tidy_data_unmemoised()} and \code{memoise()} for caching the output
}
\keyword{internal}
