% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ots_create_tidy_data.R
\name{ots_create_tidy_data_unmemoised}
\alias{ots_create_tidy_data_unmemoised}
\title{Downloads and processes the data from the API to return a human-readable tibble (unmemoised, internal)}
\usage{
ots_create_tidy_data_unmemoised(
  years = 2018,
  reporters = "usa",
  partners = "all",
  products = "all",
  sections = "all",
  groups = "all",
  table = "yr",
  max_attempts = 5,
  include_shortnames = FALSE,
  include_sections = FALSE,
  use_localhost = FALSE
)
}
\description{
A separation of \code{ots_create_tidy_data()} for making caching optional.
}
\keyword{internal}
