% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatistics-package.R
\docType{data}
\name{ots_gdp_deflator}
\alias{ots_gdp_deflator}
\title{GDP Deflator}
\format{
A data frame with 7,238 observations on the following 4 variables
\itemize{
  \item{\code{country_iso}}{ISO code of the country (e.g. "chl" means Chile)}
  \item{\code{from}}{Integer values in the range 1980-2018}
  \item{\code{to}}{Integer values in the range 1981-2019}
  \item{\code{gdp_deflator}}{Numeric value expressed as one plus 1-year deflator}
}
}
\source{
Open Trade Statistics
}
\usage{
ots_gdp_deflator
}
\description{
Year to year GDP deflator some of the countries in the OTS database. For 
countries not available in the World Bank database, rows labelled as "wld" 
are provided, which were computed as the weighted median for each year using 
the GDP of listed countries for each year expressed as constant dollars of 
the year 2010.
}
\keyword{datasets}
