% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ots_read_from_api.R
\name{ots_read_from_api}
\alias{ots_read_from_api}
\title{Reads data from the API (internal function)}
\usage{
ots_read_from_api(
  year = NULL,
  reporter_iso = NULL,
  partner_iso = NULL,
  commodity_code = "all",
  section_code = "all",
  table = "yr",
  max_attempts = 5
)
}
\description{
Accesses \code{api.tradestatistics.io} and
performs different API calls to return \code{data.frames} by reading 
\code{JSON} data. The parameters here are passed from 
\code{ots_create_tidy_data}.
}
\keyword{internal}
