\name{ots_sections}
\alias{ots_sections}
\title{A table of sections codes, names and colours associated to products}
\docType{data}
\description{
Provides section names from the Harmonized System.
}
\usage{ots_sections}
\format{
  A data frame with 1222 observations on the following 2 variables.
  \describe{
    \item{\code{product_code}}{Product code (e.g. 0101)}
    \item{\code{section_code}}{Section code (e.g. 01)}
  }
}
\examples{
ots_sections
}
\keyword{datasets}
