% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatistics-package.R
\docType{data}
\name{ots_sections}
\alias{ots_sections}
\title{OTS Sections}
\format{
A data frame with 22 rows and 2 variables
\describe{
  \item{\code{section_code}}{HS section code (e.g. '01')}
  \item{\code{section_fullname_english}}{HS section name (e.g. 'Live animals and animal products')}
}
}
\source{
Adapted from UN COMTRADE
}
\usage{
ots_sections
}
\description{
Official section names from the Harmonized System rev 2012 (HS12).
}
\keyword{datasets}
