\name{ots_sections_colors}
\alias{ots_sections_colors}
\title{A table of sections codes and hex colours associated to products}
\docType{data}
\description{
Provides section names from the Harmonized System.
}
\usage{ots_sections_colors}
\format{
  A data frame with 2 observations on the following 2 variables.
  \describe{
    \item{\code{section_code}}{Section code (e.g. 01)}
    \item{\code{section_color}}{Colour hex code (e.g. #74c0e2)}
  }
}
\examples{
ots_sections_colors
}
\keyword{datasets}
