% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatistics-package.R
\docType{data}
\name{ots_sections_colors}
\alias{ots_sections_colors}
\title{OTS Sections Colors}
\format{
A data frame with 22 rows and 2 variables
\itemize{
  \item{\code{section_code}}{Section code (e.g. '01')}
  \item{\code{section_color}}{Section hex color (e.g. '#74c0e2')}
}
}
\source{
Open Trade Statistics
}
\usage{
ots_sections_colors
}
\description{
A table of official unofficial colors to ease visualization for the sections
the Harmonized System rev 2012 (HS12, also known as H4).
}
\keyword{datasets}
