% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatistics-package.R
\docType{package}
\name{tradestatistics-package}
\alias{tradestatistics}
\alias{tradestatistics-package}
\title{tradestatistics: Open Trade Statistics API Wrapper and Utility Program}
\description{
Access 'Open Trade Statistics' API from R to download international trade data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/tradestatistics/}
  \item Report bugs at \url{https://github.com/ropensci/tradestatistics/issues/}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas \email{mavargas11@uc.cl} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Joshua Kunst (contributed to different parts of the pre-release code) [contributor]
  \item Alexey Kravchenko (reviewed 2021 version of the API) [contributor]
  \item Emma Mendelsohn (updated the functions to take available years from the API instead of hardcoded values) [contributor]
  \item Daniela de los Santos (proposed improvements to default parameters) [contributor]
  \item Emily Riederer (reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/274) [reviewer]
  \item Mark Padgham (reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/274) [reviewer]
  \item Amanda Dobbyn (reviewed a previous package that evolved into the current package for rOpenSci, see https://github.com/ropensci/onboarding/issues/217) [reviewer]
  \item Jorge Cimentada (reviewed a previous package that evolved into the current package for rOpenSci, see https://github.com/ropensci/onboarding/issues/217) [reviewer]
  \item  UN Comtrade [data contributor]
  \item  The World Bank [data contributor]
}

}
\keyword{internal}
