% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatistics-package.R
\docType{package}
\name{tradestatistics-package}
\alias{tradestatistics}
\alias{tradestatistics-package}
\title{tradestatistics: Open Trade Statistics API Wrapper and Utility Program}
\description{
Access Open Trade Statistics API from R to download international trade data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tradestatistics.github.io/tradestatistics}
  \item Report bugs at \url{https://github.com/tradestatistics/tradestatistics/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas S. \email{mvargas@dcc.uchile.cl} [copyright holder]

Other contributors:
\itemize{
  \item Joshua Kunst [contributor]
  \item Amanda Dobbyn [contributor]
  \item Jorge Cimentada [contributor]
  \item  UN Comtrade [data contributor]
}

}
\keyword{internal}
