% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatistics-package.R
\docType{package}
\name{tradestatistics-package}
\alias{tradestatistics}
\alias{tradestatistics-package}
\title{tradestatistics: Open Trade Statistics API Wrapper and Utility Program}
\description{
Access 'Open Trade Statistics' API from R to download international trade data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/tradestatistics/}
  \item Report bugs at \url{https://github.com/ropensci/tradestatistics/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas \email{mvargas@dcc.uchile.cl} (0000-0003-1017-7574) [copyright holder]

Other contributors:
\itemize{
  \item Joshua Kunst (contributed to different parts of the pre-release code) [contributor]
  \item Emily Riederer (reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/274) [reviewer]
  \item Mark Padgham (reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/274) [reviewer]
  \item Amanda Dobbyn (reviewed an archived package that evolved into the current package for rOpenSci, see https://github.com/ropensci/onboarding/issues/217) [reviewer]
  \item Jorge Cimentada (reviewed an archived package that evolved into the current package for rOpenSci, see https://github.com/ropensci/onboarding/issues/217) [reviewer]
  \item  UN Comtrade [data contributor]
  \item  Center for International Development at Harvard University [data contributor]
  \item  The Observatory of Economic Complexity [data contributor]
}

}
\keyword{internal}
